Index.js
- Added encryption

Note: Do not delete any character in the "database.json.itzporium", It's encrypted, also missing one character will affect all data.

+++ index.js
@@ -1,10 +1,19 @@
 // The password
 const authToken = "the token here or env token"
+const accessJSON_key = "this is the key, we gonna use to encrypt and decrypt data and save to a file, you can set this with authToken too";
+
+// Create database file, if not found
+if (!require("fs").existsSync("./database.json.itzporium")) {
+    require("fs").writeFileSync("./database.json.itzporium", require("crypto-js").AES.encrypt("[]", accessJSON_key?.toString())?.toString());
+}
 
 // Database Variable
 const express = require("express");
+const cryptojs = require("crypto-js");
 const fs = require("fs");
-let db = require("./database.json");
+
+let db_require = fs.readFileSync("./database.json.itzporium", "utf-8");
+let db = JSON.parse(cryptojs.AES.decrypt(db_require, accessJSON_key?.toString()).toString(cryptojs.enc.Utf8));
 
 let localdb = new Map(
     db.map(obj => {
@@ -38,13 +47,12 @@ app.use(express.json());
     If info is not vaild, it will return `undefined`.
 */
 
-
 // Error handler
-process.on('uncaughtException', function() { /*console.log("[expressDB]: error found!")*/ })
+process.on("uncaughtException", function() { /*console.log("[expressDB]: error found!")*/ })
 
 // Bad json format
 app.use((err, req, res, next) => {
-    if (err instanceof SyntaxError && err.status === 400 && 'body' in err) {
+    if (err instanceof SyntaxError && err.status === 400 && "body" in err) {
         return res.status(400).json({
             status: false,
             keyname: null,
@@ -56,7 +64,7 @@ app.use((err, req, res, next) => {
 
 // Authentication
 app.use((req, res, next) => {
-    if (req instanceof SyntaxError && req.status === 400 && 'body' in req) {
+    if (req instanceof SyntaxError && req.status === 400 && "body" in req) {
         return res.status(400).json({
             status: false,
             keyname: null,
@@ -92,7 +100,6 @@ app.use((req, res, next) => {
 app.post("/get", (req, res) => {
     try {
         // check if the db is exits
-        const execpath = `./database.json`
         const decodeinfo = Array.from(localdb).map(([key, value]) => ({
             key,
             value,
@@ -117,36 +124,35 @@ app.post("/get", (req, res) => {
 app.post("/set", (req, res) => {
     try {
         // check if the db is exits
-        const execpath = `./database.json`
-        let decodeinfo = db
+        if (fs.existsSync("./database.json.itzporium")) {
+            try {
+                let test1 = req.body["key"];
+                let test2 = req.body["value"];
 
-        try {
-            let test1 = req.body["key"]
-            let test2 = req.body["value"];
-        } catch (err) {
-            res.json({
-                status: false,
-                keyname: null,
-                data: null,
-                message: "failed",
-            })
+                localdb.set(req.body["key"], req.body["value"])
+    
+                const arr = Array.from(localdb).map(([key, value]) => ({
+                    key,
+                    value,
+                }));
+        
+                fs.writeFileSync("./database.json.itzporium", cryptojs.AES.encrypt(JSON.stringify(arr), accessJSON_key?.toString())?.toString());
+        
+                res.json({
+                    status: true,
+                    keyname: req.body["key"],
+                    data: req.body["value"],
+                    message: "sucessfully",
+                })
+            } catch (err) {
+                res.json({
+                    status: false,
+                    keyname: null,
+                    data: null,
+                    message: "failed",
+                })
+            }
         }
-
-        localdb.set(req.body["key"], req.body["value"])
-
-        const arr = Array.from(localdb).map(([key, value]) => ({
-            key,
-            value,
-        }));
-
-        fs.writeFileSync('./database.json', JSON.stringify(arr));
-
-        res.json({
-            status: true,
-            keyname: req.body["key"],
-            data: req.body["value"],
-            message: "sucessfully",
-        })
     } catch {
         res.json({
             status: false,
@@ -160,40 +166,39 @@ app.post("/set", (req, res) => {
 app.delete("/del", (req, res) => {
     try {
         // check if the db is exits
-        const execpath = `./database.json`
-        let decodeinfo = db;
+        if (fs.existsSync("./database.json.itzporium")) {
+            try {
+                let test1 = req.body["key"];
 
-        try {
-            let test1 = req.body["key"]
-        } catch (err) {
-            res.json({
-                status: false,
-                keyname: null,
-                data: null,
-                message: "failed",
-            })
+                let localdbdata = null;
+    
+                if (localdb.has(req.body["key"])) {
+                    localdbdata = localdb.get(req.body["key"])
+                    localdb.delete(req.body["key"]);
+                }
+        
+                const arr = Array.from(localdb).map(([key, value]) => ({
+                    key,
+                    value,
+                }));
+        
+                fs.writeFileSync("./database.json.itzporium", cryptojs.AES.encrypt(JSON.stringify(arr), accessJSON_key?.toString())?.toString());
+        
+                res.json({
+                    status: true,
+                    keyname: req.body["key"] || null,
+                    data: localdbdata || null,
+                    message: "sucessfully",
+                })
+            } catch (err) {
+                res.json({
+                    status: false,
+                    keyname: null,
+                    data: null,
+                    message: "failed",
+                })
+            }
         }
-
-        let localdbdata = null;
-
-        if (localdb.has(req.body["key"])) {
-            localdbdata = localdb.get(req.body["key"])
-            localdb.delete(req.body["key"]);
-        }
-
-        const arr = Array.from(localdb).map(([key, value]) => ({
-            key,
-            value,
-        }));
-
-        fs.writeFileSync('./database.json', JSON.stringify(arr));
-
-        res.json({
-            status: true,
-            keyname: req.body["key"] || null,
-            data: localdbdata || null,
-            message: "sucessfully",
-        })
     } catch {
         res.json({
             status: false,
@@ -207,40 +212,39 @@ app.delete("/del", (req, res) => {
 app.post("/has", (req, res) => {
     try {
         // check if the db is exits
-        const execpath = `./database.json`
-        let decodeinfo = db;
-
-        try {
-            let test1 = req.body["key"]
-        } catch (err) {
-            res.json({
-                status: false,
-                keyname: null,
-                data: null,
-                message: "failed",
-            })
-        }
-
-        let localdbdata = null;
-
-        if (localdb.has(req.body["key"])) {
-            localdbdata = localdb.get(req.body["key"])
-        }
-
-        if (localdb.has(req.body["key"])) {
-            res.json({
-                status: true,
-                keyname: req.body["key"] || null,
-                data: localdbdata || null,
-                message: "sucessfully",
-            })
-        } else {
-            res.json({
-                status: false,
-                keyname: null,
-                data: null,
-                message: "failed",
-            })
+        if (fs.existsSync("./database.json.itzporium")) {
+            try {
+                let test1 = req.body["key"]
+            } catch (err) {
+                res.json({
+                    status: false,
+                    keyname: null,
+                    data: null,
+                    message: "failed",
+                })
+            }
+    
+            let localdbdata = null;
+    
+            if (localdb.has(req.body["key"])) {
+                localdbdata = localdb.get(req.body["key"])
+            }
+    
+            if (localdb.has(req.body["key"])) {
+                res.json({
+                    status: true,
+                    keyname: req.body["key"] || null,
+                    data: localdbdata || null,
+                    message: "sucessfully",
+                })
+            } else {
+                res.json({
+                    status: false,
+                    keyname: null,
+                    data: null,
+                    message: "failed",
+                })
+            }
         }
     } catch {
         res.json({
@@ -255,18 +259,17 @@ app.post("/has", (req, res) => {
 app.post("/clear", (req, res) => {
     try {
         // check if the db is exits
-        const execpath = `./database.json`
-        let decodeinfo = db;
+        if (fs.existsSync("./database.json.itzporium")) {
+            localdb.clear(); // clear everything have in it local database
+            fs.writeFileSync("./database.json.itzporium", cryptojs.AES.encrypt("[]", accessJSON_key?.toString())?.toString()); // clear the stored json database
 
-        localdb.clear(); // clear everything have in it local database
-        fs.writeFileSync("./database.json", "[]"); // clear the stored json database
-
-        res.json({
-            status: true,
-            keyname: [],
-            data: [],
-            message: "sucessfully",
-        })
+            res.json({
+                status: true,
+                keyname: [],
+                data: [],
+                message: "sucessfully",
+            })
+        }
     } catch {
         res.json({
             status: false,
